public 설정
