--- api/admins/serializers.py
@@ -1,9 +1,0 @@
-from rest_framework import serializers
-
-from api.publics.models import User_APT
-
-class AccountSerializer(serializers.ModelSerializer):
-    
-    class Meta:
-        model = User_APT
-        fields = ('floor','cost',)
