+++ api/publics/urls.py
@@ -1,0 +1,10 @@
+from django.urls import path, include
+
+from rest_framework.routers import DefaultRouter
+from .views                 import AccountViewSet
+
+router = DefaultRouter()
+router.register('', AccountViewSet)
+urlpatterns = [
+    path('', include(router.urls)),
+]
